Created api directory and created server file to add endpoints
