+++ routes/users-route.js
@@ -4,3 +4,4 @@ const users = require('../users/users-model.js')
 
 const router = express.Router()
 
+router.get('/')
